Fix component export issues for build deployment
- Add default exports to all component index files that were missing them
- Fix StarRating, InlineRating, and all review components exports
- Fix UI components (InlineEditableField, Newsletter, LevelProgressBar) exports
- Fix admin components (ArcModal, BookModal, CharacterModal, etc.) exports
- Fix shop components (BookStore, BookNavigator, GridView, etc.) exports
- Add placeholder rate limiting and validation functions to prevent build errors

This resolves the 'does not contain a default export' errors in Vercel deployment
