+++ package.json
@@ -31,6 +31,7 @@
     "@headlessui/react": "^2.2.6",
     "@heroicons/react": "^2.2.0",
     "@hookform/resolvers": "^5.2.0",
+    "@storybook/nextjs": "^9.0.18",
     "@supabase/ssr": "^0.6.1",
     "@supabase/supabase-js": "^2.52.1",
     "@tailwindcss/typography": "^0.5.16",
@@ -39,8 +40,8 @@
     "framer-motion": "^12.23.12",
     "lucide-react": "^0.534.0",
     "next": "^13.4.19",
-    "react": "19.1.0",
-    "react-dom": "19.1.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-force-graph-2d": "^1.28.0",
     "react-hook-form": "^7.61.1",
     "react-hot-toast": "^2.5.2",
@@ -57,7 +58,7 @@
     "@playwright/test": "^1.54.1",
     "@storybook/addon-essentials": "^8.6.14",
     "@storybook/addon-interactions": "^8.6.14",
-    "@storybook/addon-links": "^9.0.18",
+    "@storybook/addon-links": "^8.6.14",
     "@storybook/react": "^9.0.18",
     "@storybook/react-vite": "^9.0.18",
     "@storybook/test": "^8.6.14",
@@ -75,7 +76,7 @@
     "jest": "^30.0.5",
     "jest-environment-jsdom": "^30.0.5",
     "msw": "^2.10.4",
-    "storybook": "^8.6.14",
+    "storybook": "^9.0.18",
     "tailwindcss": "^4",
     "typescript": "^5",
     "whatwg-fetch": "^3.6.20"
