+++ src/components/features/admin/ArcManager/ArcManager.tsx
@@ -1,7 +1,7 @@
 'use client'
 import { useState, useEffect } from 'react'
 import { PlusIcon, PencilIcon, TrashIcon } from '@heroicons/react/24/outline'
-import ArcModal from '@/components/features/admin/ArcModal'
+import { ArcModal } from '@/components/features/admin/ArcModal'
 
 interface Arc {
   id: number
