+++ src/components/features/admin/ArcModal/index.ts
@@ -1,1 +1,2 @@
-﻿export { default as ArcModal } from './ArcModal';
+export { default } from './ArcModal';
+export { default as ArcModal } from './ArcModal';
