+++ src/components/features/admin/CharacterModal/index.ts
@@ -1,1 +1,2 @@
-﻿export { default as CharacterModal } from './CharacterModal';
+export { default } from './CharacterModal';
+export { default as CharacterModal } from './CharacterModal';
