+++ src/components/features/admin/EasterEgg/index.ts
@@ -1,1 +1,2 @@
-﻿export { default as EasterEgg } from './EasterEgg';
+export { default } from './EasterEgg';
+export { default as EasterEgg } from './EasterEgg';
