+++ src/components/features/admin/EasterEggAdminToggle/index.ts
@@ -1,1 +1,2 @@
-﻿export { default as EasterEggAdminToggle } from './EasterEggAdminToggle';
+export { default } from './EasterEggAdminToggle';
+export { default as EasterEggAdminToggle } from './EasterEggAdminToggle';
