+++ src/components/features/admin/IssueModal/index.ts
@@ -1,1 +1,2 @@
-﻿export { default as IssueModal } from './IssueModal';
+export { default } from './IssueModal';
+export { default as IssueModal } from './IssueModal';
