+++ src/components/features/reviews/ReviewForm/index.ts
@@ -1,1 +1,2 @@
-﻿export { default as ReviewForm } from './ReviewForm';
+export { default } from './ReviewForm';
+export { default as ReviewForm } from './ReviewForm';
