+++ src/components/features/reviews/ReviewList/ReviewList.tsx
@@ -2,8 +2,8 @@ import React, { useState } from 'react';
 import { Review } from '@/lib/types';
 import { useAuth } from '@/app/contexts/AuthContext';
 import { reviewsApi } from '@/lib/hooks/useReviews';
-import StarRating from '../StarRating';
-import ReviewForm from '../ReviewForm';
+import { StarRating } from '../StarRating';
+import { ReviewForm } from '../ReviewForm';
 
 interface ReviewListProps {
   reviews: Review[];
