+++ src/components/features/reviews/ReviewList/index.ts
@@ -1,1 +1,2 @@
-﻿export { default as ReviewList } from './ReviewList';
+export { default } from './ReviewList';
+export { default as ReviewList } from './ReviewList';
