+++ src/components/features/shop/BookNavigator/index.ts
@@ -1,1 +1,2 @@
-﻿export { default as BookNavigator } from './BookNavigator';
+export { default } from './BookNavigator';
+export { default as BookNavigator } from './BookNavigator';
