+++ src/components/features/shop/GridView/index.ts
@@ -1,1 +1,2 @@
-﻿export { default as GridView } from './GridView';
+export { default } from './GridView';
+export { default as GridView } from './GridView';
