+++ src/components/ui/InlineEditableField/index.ts
@@ -1,1 +1,2 @@
-﻿export { default as InlineEditableField } from './InlineEditableField';
+export { default } from './InlineEditableField';
+export { default as InlineEditableField } from './InlineEditableField';
