+++ src/components/ui/LevelProgressBar/index.ts
@@ -1,0 +1,2 @@
+export { default } from './LevelProgressBar';
+export { default as LevelProgressBar } from './LevelProgressBar';
