+++ src/components/ui/Newsletter/index.ts
@@ -1,1 +1,2 @@
-﻿export { default as Newsletter } from './Newsletter';
+export { default } from './Newsletter';
+export { default as Newsletter } from './Newsletter';
