+++ src/lib/rateLimit.ts
@@ -1,3 +1,8 @@
 export const rateLimit = {
     // Add your rate limiting logic here
 };
+
+// FIXME: Replace with real rate limiting logic
+export const applyGeneralRateLimit = () => true;
+export const applyProgressRateLimit = () => true;
+export const applyReviewRateLimit = () => true;
